Initial merak compute
