+++ api/proto/v1/topology.proto
@@ -1,3 +1,3 @@
 syntax="proto3";
 package merak;
-option go_package = "/";
+option go_package = "merak/";
