--- services/README.md
@@ -1,1 +1,0 @@
-Folder for the source code of each microservices.
