+++ src/README.md
@@ -1,0 +1,1 @@
+Folder for the source code of each microservices.
