+++ src/common/constants.go
@@ -1,0 +1,6 @@
+package common
+
+const (
+	TEMPORAL_PORT = "7233"
+	TEMPORAL_ENV  = "TEMPORAL"
+)
