+++ src/module.mk
@@ -1,0 +1,6 @@
+module := src
+
+submodules := merak-compute
+-include $(patsubst %, $(module)/%/module.mk, $(submodules))
+
+all:: $(submodules)
